--- app/src/main/java/com/csb/koreamoviedb_mvvm/model/network/ApiClient.kt
@@ -1,37 +1,0 @@
-package com.csb.koreamoviedb_mvvm.model.network
-
-import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
-import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.json.Json
-import okhttp3.MediaType.Companion.toMediaType
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-
-object ApiClient {
-    private val mediaType = "application/json; charset=UTF-8".toMediaType()
-    private val json = Json { ignoreUnknownKeys = true }
-
-    private const val BASEURL = "http://api.koreafilm.or.kr"
-
-    private val loggingInterceptor = HttpLoggingInterceptor().apply {
-        level = HttpLoggingInterceptor.Level.BODY
-    }
-
-    private val okHttpClient = OkHttpClient.Builder()
-        .addInterceptor(loggingInterceptor)
-        .build()
-
-    @OptIn(ExperimentalSerializationApi::class)
-    val client: Retrofit by lazy {
-        Retrofit.Builder()
-            .baseUrl(BASEURL)
-            .client(okHttpClient)
-            .addConverterFactory(json.asConverterFactory(mediaType))
-            .build()
-    }
-
-    val movieApiService: MovieApiService by lazy {
-        client.create(MovieApiService::class.java)
-    }
-}
