+++ app/src/main/java/com/csb/koreamoviedb_mvvm/presentation/search/component/GlideBox.kt
@@ -1,0 +1,35 @@
+package com.csb.koreamoviedb_mvvm.presentation.search.component
+
+
+import android.util.Log
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
+import com.bumptech.glide.integration.compose.GlideImage
+
+
+@OptIn(ExperimentalGlideComposeApi::class)
+@Composable
+fun GlideBox(
+    model: String,
+    modifier: Modifier = Modifier
+) {
+    Log.d("GlideBox",model)
+    Box(
+        modifier = modifier
+            .fillMaxSize()
+    ) {
+        GlideImage(
+            modifier = Modifier
+                .fillMaxSize()
+                .align(Alignment.Center),
+            model = model,
+            contentDescription = "포스터이미지",
+            contentScale = ContentScale.Fit
+        )
+    }
+}
