+++ app/src/main/java/com/csb/koreamoviedb_mvvm/repository/MovieRepository.kt
@@ -1,0 +1,12 @@
+package com.csb.koreamoviedb_mvvm.repository
+
+import com.csb.koreamoviedb_mvvm.domain.ResultMovieClass
+
+interface MovieRepository {
+    suspend fun getMovies(
+        query: String? = null,
+        title: String? = null,
+        actor: String? = null,
+        director: String? = null
+    ): List<ResultMovieClass>
+}
