--- app/src/main/java/com/csb/koreamoviedb_mvvm/view/component/CustomProgressDialog.kt
@@ -1,43 +1,0 @@
-package com.csb.koreamoviedb_mvvm.view.component
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Dialog
-
-@Composable
-fun CustomProgressDialog(
-    isShowing: Boolean,
-    onDismissRequest: () -> Unit = {},
-) {
-    if (isShowing) {
-        Dialog(onDismissRequest = onDismissRequest) {
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .wrapContentHeight()
-                    .background(Color.Transparent),
-                contentAlignment = Alignment.Center
-            ) {
-                Column(
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    CircularProgressIndicator(
-                        modifier = Modifier.size(60.dp),
-                        color = Color.Black,
-                        strokeWidth = 6.dp
-                    )
-                }
-            }
-        }
-    }
-}
