--- app/src/main/java/com/csb/koreamoviedb_mvvm/view/component/MovieItem.kt
@@ -1,126 +1,0 @@
-package com.csb.koreamoviedb_mvvm.view.component
-
-import android.util.Log
-import androidx.activity.ComponentActivity
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
-
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.rememberNavController
-
-import com.csb.koreamoviedb_mvvm.model.data.ResultMovieClass
-import com.csb.koreamoviedb_mvvm.tools.RootScreen
-import com.csb.koreamoviedb_mvvm.viewmodel.SharedMovieViewModel
-
-@Composable
-fun MovieItem(
-    navController: NavHostController,
-    movie: ResultMovieClass
-) {
-
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(12.dp)
-            .clickable {
-                navController.currentBackStackEntry?.savedStateHandle?.set("movie",movie)
-                navController.navigate(RootScreen.SCREEN_DETAIL.name)
-            }
-    ) {
-        Column(
-            modifier = Modifier
-                .height(100.dp)
-                .width(100.dp)
-        ) {
-            val posterUrl = movie.posters.split("|").firstOrNull()?.trim()
-            Log.d("posterUrl",posterUrl.toString())
-            GlideBox(
-                model = posterUrl ?: "",
-                modifier = Modifier.fillMaxSize()
-            )
-        }
-
-        Spacer(modifier = Modifier.width(5.dp))
-
-        Column(modifier = Modifier.align(Alignment.CenterVertically)) {
-            Text(
-                text = movie.title,
-                style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.Bold)
-            )
-            Spacer(modifier = Modifier.height(4.dp))
-            Text("상영시간: ${movie.runtime}분", style = MaterialTheme.typography.bodyMedium)
-            Text("관람등급: ${movie.ratingGrade}", style = MaterialTheme.typography.bodyMedium)
-        }
-    }
-}
-
-@Composable
-fun MovieItemPreviewOnly(movie: ResultMovieClass) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(12.dp)
-    ) {
-        Column(
-            modifier = Modifier
-                .height(100.dp)
-                .width(100.dp)
-        ) {
-            val posterUrl = movie.posters.split("|").firstOrNull()?.trim()
-            Log.d("posterUrl", posterUrl.toString())
-            GlideBox(
-                model = posterUrl ?: "",
-                modifier = Modifier.fillMaxSize()
-            )
-        }
-
-        Spacer(modifier = Modifier.width(5.dp))
-
-        Column(modifier = Modifier.align(Alignment.CenterVertically)) {
-            Text(
-                text = movie.title,
-                style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.Bold)
-            )
-            Spacer(modifier = Modifier.height(4.dp))
-            Text("상영시간: ${movie.runtime}분", style = MaterialTheme.typography.bodyMedium)
-            Text("관람등급: ${movie.ratingGrade}", style = MaterialTheme.typography.bodyMedium)
-        }
-    }
-}
-
-
-@Preview(showBackground = true)
-@Composable
-fun PreviewMovieItem() {
-    val dummyMovie = ResultMovieClass(
-        title = "헌트",
-        directors = listOf("이정재"),
-        actors = listOf("이정재", "정우성", "전혜진", "허성태", "김종수"),
-        plots = listOf("ㅈㄱㄹ 국문", "ㅈㄱㄹ 영문"),
-        detailInfo = "https://movie.daum.net/moviedb/main?movieId=123456",
-        posters = "http://file.koreafilm.or.kr/thm/02/00/03/51/tn_DPF05297A.jpg",
-        runtime = "125",
-        ratingGrade = "15세이상관람가",
-        releaseDate = "20220810"
-    )
-
-    MovieItemPreviewOnly(movie = dummyMovie)
-}
-
