--- app/src/main/java/com/csb/koreamoviedb_mvvm/viewmodel/SharedUiState.kt
@@ -1,9 +1,0 @@
-package com.csb.koreamoviedb_mvvm.viewmodel
-
-import javax.inject.Inject
-import javax.inject.Singleton
-
-@Singleton
-class SharedUiState@Inject constructor()  {
-
-}
